% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlify_barchart_horizontal.R
\name{plotlify_barchart_horizontal}
\alias{plotlify_barchart_horizontal}
\title{Turn a static horizontal bar chart into an interactive plotly object.}
\usage{
plotlify_barchart_horizontal(ggplot_input)
}
\arguments{
\item{ggplot_input}{static ggplot horizontal bar chart.}
}
\value{
An interactive plotly chart.
}
\description{
Turn a static horizontal bar chart into an interactive plotly object.
}
\examples{
NULL
}
