% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlify_choromap_univar.R
\name{plotlify_choromap_univar}
\alias{plotlify_choromap_univar}
\title{Turn a static univariate choropleth map into an interactive plotly object.}
\usage{
plotlify_choromap_univar(ggplot_input)
}
\arguments{
\item{ggplot_input}{static ggplot chart with a univariate choropleth map}
}
\value{
An interactive plotly chart.
}
\description{
Turn a static univariate choropleth map into an interactive plotly object.
}
\examples{
NULL
}
