% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlify_scatter.R
\name{plotlify_scatter}
\alias{plotlify_scatter}
\title{Turn a static scatterplot into an interactive plotly object.}
\usage{
plotlify_scatter(ggplot_input)
}
\arguments{
\item{ggplot_input}{static ggplot scatterplot}
}
\value{
An interactive plotly chart.
}
\description{
Turn a static scatterplot into an interactive plotly object.
}
