% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlify_single_time_series.R
\name{plotlify_single_time_series}
\alias{plotlify_single_time_series}
\title{Turn a static single time-series chart into an interactive plotly object.}
\usage{
plotlify_single_time_series(ggplot_input)
}
\arguments{
\item{ggplot_input}{static ggplot chart with a single timeseries}
}
\value{
An interactive plotly chart.
}
\description{
Turn a static single time-series chart into an interactive plotly object.
}
