% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlify_timeseries_multivar.R
\name{plotlify_timeseries_multivar}
\alias{plotlify_timeseries_multivar}
\title{Turn a static multivariate time-series chart into an interactive plotly object.}
\usage{
plotlify_timeseries_multivar(ggplot_input)
}
\arguments{
\item{ggplot_input}{static ggplot chart with multiple timeseries}
}
\value{
An interactive plotly chart.
}
\description{
Turn a static multivariate time-series chart into an interactive plotly object.
}
\examples{
NULL
}
