% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlify_timeseries_univar.R
\name{plotlify_timeseries_univar}
\alias{plotlify_timeseries_univar}
\title{Turn a static univariate time-series chart into an interactive plotly object.}
\usage{
plotlify_timeseries_univar(ggplot_input)
}
\arguments{
\item{ggplot_input}{static ggplot chart with a single timeseries}
}
\value{
An interactive plotly chart.
}
\description{
Turn a static univariate time-series chart into an interactive plotly object.
}
\examples{
NULL
}
