% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gov.R
\name{theme_gov}
\alias{theme_gov}
\title{Custom ggplot theme for most static charts}
\usage{
theme_gov()
}
\description{
To be used with most chart types other than horizontal bar charts and maps.
}
\examples{
NULL
}
