% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gov_hbar.R
\name{theme_gov_hbar}
\alias{theme_gov_hbar}
\title{Custom ggplot theme for static horizontal bar charts}
\usage{
theme_gov_hbar()
}
\description{
To be used with horizontal bar charts only.
}
\examples{
NULL
}
